Sepearated dealership search details/edit container logic from presentation
